requirements.txt
1streamlit>=1.32.0
pandas>=2.2.0
requests>=2.31.0
folium>=0.16.0
streamlit-folium>=0.18.0
scikit-learn>=1.4.0
feedparser>=6.0.11
plotly>=5.20.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
